Change requestResponseMessage identifier to a number (#8)
Every request dispatched over the postMessage boundary is tagged with an
identifier. The promise for the request is stored in a map keyed by this
ID. When the remote end responds to the request, the dispatch layer
looks up the promise to resolve by this ID.

Prior to this change, the dispatch logic would assign random string ids.
This meant that each RPC dispatch would need to:
1. Generate a random integer
2. Convert it to a string
3. Transfer the string over the worker boundary
4. Transfer the string back over the worker boundary when responding to
the RPC

These identifiers are internal and there is no need for them to be
random or strings. We can save on all of the above work by using an
incrementing counter.