Change Message and Wire type enums to numbers (#11)
These enums are internal values for indicating the type of a message or
wire value. We can avoid transferring a string by changing them to
numbers. The typescript code is unchanged but the produced javascript
uses numbers rather than strings which saves on transfer cost, string
initialization/copy cost, and comparison cost with no loss in
readability of the source.